+++ data/DoctrineORMModule/Proxy/__CG__DtlEmployeeEntityEmployee.php
@@ -179,7 +179,7 @@ class Employee extends \DtlEmployee\Entity\Employee implements \Doctrine\ORM\Pro
     public function getId()
     {
         if ($this->__isInitialized__ === false) {
-            return  parent::getId();
+            return (int)  parent::getId();
         }
