+++ data/DoctrineORMModule/Proxy/__CG__DtlSellerEntitySeller.php
@@ -179,7 +179,7 @@ class Seller extends \DtlSeller\Entity\Seller implements \Doctrine\ORM\Proxy\Pro
     public function getId()
     {
         if ($this->__isInitialized__ === false) {
-            return  parent::getId();
+            return (int)  parent::getId();
         }
