+++ data/DoctrineORMModule/Proxy/__CG__DtlShopmanEntityShopman.php
@@ -179,7 +179,7 @@ class Shopman extends \DtlShopman\Entity\Shopman implements \Doctrine\ORM\Proxy\
     public function getId()
     {
         if ($this->__isInitialized__ === false) {
-            return  parent::getId();
+            return (int)  parent::getId();
         }
